requirements_complete.txt
# Web framework and server
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0

# Google Cloud AI services
google-cloud-aiplatform==1.38.0
google-generativeai==0.3.2

# Vector database and embeddings
chromadb==0.4.15
sentence-transformers==2.2.2

# Data processing
pandas==2.0.3
numpy==1.24.3

# Google services integration
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0

# HTTP and utilities
httpx==0.25.2
requests==2.31.0
python-multipart==0.0.6

# File processing
openpyxl==3.1.2
jsonlines==3.1.0
Pillow==9.5.0

# Additional dependencies
pydantic==2.5.0 